+++ module-apps/application-alarm-clock/windows/NewEditAlarmWindow.cpp
@@ -3,6 +3,7 @@
 
 #include "NewEditAlarmWindow.hpp"
 #include "application-alarm-clock/data/AlarmsData.hpp"
+#include "application-calendar/data/OptionParser.hpp"
 #include <module-db/Interface/AlarmsRecord.hpp>
 
 namespace app::alarmClock
@@ -52,6 +53,13 @@ namespace app::alarmClock
                 alarmRecord = rec->getData();
             }
             presenter->loadData(alarmRecord);
+        }
+
+        if (mode == gui::ShowMode::GUI_SHOW_RETURN) {
+            if (auto receivedData = dynamic_cast<WeekDaysRepeatData *>(data); receivedData != nullptr) {
+                presenter->updateRepeat(alarmRecord, *receivedData);
+                presenter->loadRepeat(alarmRecord);
+            }
         }
     }